+++ src/reflection/ApplicationPropertiesClassReflectionExtension.php
@@ -13,8 +13,8 @@ use PHPStan\Reflection\{
 };
 use PHPStan\Reflection\Annotations\AnnotationsPropertiesClassReflectionExtension;
 use PHPStan\Reflection\Dummy\DummyPropertyReflection;
-use PHPStan\Type\ObjectType;
 use PHPStan\Type\Generic\GenericObjectType;
+use PHPStan\Type\ObjectType;
 use yii2\extensions\phpstan\ServiceMap;
 
 /**
@@ -99,10 +99,11 @@ final class ApplicationPropertiesClassReflectionExtension implements PropertiesC
             );
         }
 
-        if ($this->serviceMap->isUserComponentClass($propertyName)) {
+        $identityClass = $this->serviceMap->getUserComponentClassById($propertyName);
+        if (!is_null($identityClass)) {
             return new ComponentPropertyReflection(
                 new DummyPropertyReflection($propertyName),
-                new GenericObjectType('yii\web\User', [new ObjectType($this->serviceMap->getUserComponentClassById($propertyName))]),
+                new GenericObjectType('yii\web\User', [new ObjectType($identityClass)]),
                 $classReflection,
             );
         }